+++ rust/cli/src/lib/error.rs
@@ -140,9 +140,7 @@ pub enum Error {
     #[error("Unsupported chip type '{0}'.\n  Supported types for this board: {1}")]
     UnsupportedChipType(String, String),
 
-    #[error(
-        "This board does not support chip types {1}.\n  Supported types: {2}\n  Use --allow-unsupported-chip-type to override."
-    )]
+    #[error("This board cannot serve chip types {1}.\n  Supported types: {2}")]
     UnsupportedBoardChipType(String, String, String),
 
     #[error(